Fix type mismatch

+++ MediaPlayerDemo/src/main/java/net/protyposis/android/mediaplayerdemo/VideoViewActivity.java
@@ -79,7 +79,7 @@ public class VideoViewActivity extends Activity {
         super.onRestoreInstanceState(savedInstanceState);
         mVideoUri = savedInstanceState.getParcelable("uri");
         mVideoPosition = savedInstanceState.getInt("position");
-        mVideoPlaybackSpeed = savedInstanceState.getInt("playbackSpeed");
+        mVideoPlaybackSpeed = savedInstanceState.getFloat("playbackSpeed");
         mVideoPlaying = savedInstanceState.getBoolean("playing");
     }
